A pasted selection is supposed to snap to the grid, but this cannot work the way its implemented currently. If one of the points is currently snapped, then how can we make sure that that SAME point is snapped again after pasting? If it's another point that snaps, then this is useless...
Furthermore it's also buggy currently (it messes up the grid snapping sensitivity), so I'm removing this functionality. If someone wants it back, we should rather round the paste offset to a multiple of the grid pitch. This way, something that was snapped to the grid before will stay snapped after pasting.