reintroducing code that initially committed in revision 18594 and the have been reverted on revision 18607.
* implemented handling of font-style, font-variant, font-weight, and 
font-stretch attributes on font-face tags

This have been previously reverted because I thought they were already handled in node->style, but now I see that this code is needed for font description (used in the 
font tag to declare font attributes to be stored in the user agent font database).

The node->style info is used for font specification (used by document authors with the purpose of declaring the properties of fonts that they would like in their SVG 
documents).