Sorry, I got off on a branch and ended up with a bunch of things.  I'm just going to flatten and merge...  Here's the list.
* Added a focus mode activated by Shift+F11.  The goal of this mode is to 
remove all toolbars for a short period so that you maximize screen area.  
Useful on small screens.  Also when you know lots of shortcuts.

* Created what I'm calling "Quick Zoom."  The idea here is to have a modal 
zoom for touching up something with fine detail and then returning to what 
you're doing.  This is activated with the letter "Q" on the keyboard.  When 
you release it, you return back to where you were.  This will zoom in on 
selected objects, or if you're in the node tool selected nodes.

* Added .svg on the temporary files in extensions.  This'll make many of 
them happier.  This only works on recent versions of GLib, but shouldn't 
break older ones more than they already are.

* Moved the Inkscape configuration directory on Linux from ~/.inkscape 
to ~/.config/Inkscape.  This is the new way to do things with the cross 
desktop naming spec.  I'm unsure whether we should be putting crash dumps 
in .config or .cache though.

* Removed 'tools_switch_current' because every usage of it already had a 
pointer to where it needed to go.  Removes usage of globals.

* Made it so that dialogs will be transparent when not focused.  This is 
an alternate to having the docked, and one that I like better as I feel it 
gives me more screen area.  You can adjust how much transparency and the 
speed of the animation in the preferences dialog.  (Note: this requires 
GTK+ 2.12 and a compositor, but gracefully degrades if you don't have 
either)