Fix bug 215906
The comments:
https://bugs.launchpad.net/inkscape/+bug/215906
highlight that the user ~/Library/fonts folder is not used by Inkscape, which is weird because it most likely was before. We provide our own fonts.conf and this fixes it to add ~/Library/fonts.

In addition, it currently comments the system fonts because most of them are in dfont format and are not yet handled well by pango.