-  Remove the old native gtk building script.  It is no longer needed:
The easiest way to do this now is to install a second macport tree as per the
instructions on our CompilingMacOsX wiki page then edit the
/opt/local-native/etc/macports/variants.conf file to have the line:
	+universal +no_x11 +quartz
This will cause it to build all packages with the no_x11 and quartz variants if
they exist.  You probably also need to set alternate applications_dir and
frameworks_dir variables in the /opt/local-native/etc/macports/macports.conf
file so that things don't clash with your standard macports tree.