988601-changes_2013_03_20a.patch
1.  Fixes the clang warnings noted in a post above, other than those
associated with alignment caused by casting.

2.  Fixes some minor rounding errors in both WMF and EMF input/output.
Round trip open/save cycles are conservative for EMF and WMF files 
(excluding any features that are not full supported in inkscape or the
target file format, for instance, gradients, which must be emulated.)

3.  Fixed a missing break in the input WMF LINETO record handling, which
was falling through into the MOVETO and generating a harmless extra "M"
operation in a path.

4.  WMF has no POLYPOLYLINE record.   However input that maps into
essentially a polypolyline record in SVG is common, for instance dashed
lines that have been converted to line segments.  These end up in SVG as
a series of M L M L draw commands in the path.  Earlier each M L pair
was going out as a polyline record, now they go as a series of
MOVETO/LINETO records.  The primary reason for this change is that
without this change the behavior described in (2) does not occur.

5.  Fixed an issue where polyline and polygon records in some instances
ended up with an extra copy of their last point.