Updating libdepixelize.
This update is not focused on new features, but on stability.

Bugs affecting weird image sizes (single line or single row images) were
fixed.

A bug on algorithm that could cause wrong result of heuristics on very
rare images was also fixed. The correct behaviour required storing all
votes before really doing any removal and this extra work made this step
of the algorithm 328% slower on one sample image. If the old *_safe
function was used, the slowdown could be decreased to 7.7% on the same
sample image, but current focus on Inkscape timeline is stabilization,
then I'll delay performance optimization to later. Also, the affected
step is only responsible for 18% (on the maximum case) of the
performance. The performance tests were done using new profiling code
that is disabled by default and shouldn't impact stability.