Add extension to set attributes for bitmap images (bug #1357808)
Inkscape >= 0.91 adds the attribute 'preserveAspectRatio'
to imported bitmap images, and sets it to 'none' (to support
non-uniform scaling). Unlike older versions it also respects the
attribute for rendering. This change may break Inkscape documents
with embedded or linked bitmap images which had been created with
older versions of Inkscape (the images do not have the attribute
set, and thus default to 'xMidYMid' with enforced uniform scaling).

The extension allows to add the attribute to selected or all bitmap
images in the current document (including bitmap images used as
masks, which are stored in the <defs> section).